@@ -49,31 +49,31 @@
     "access": "public"
   },
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@angular/animations": "~8.2.8",
+    "@angular/common": "~8.2.8",
+    "@angular/compiler": "~8.2.8",
+    "@angular/core": "~8.2.8",
+    "@angular/forms": "~8.2.8",
+    "@angular/platform-browser": "~8.2.8",
+    "@angular/platform-browser-dynamic": "~8.2.8",
+    "@angular/router": "~8.2.8",
     "@datorama/akita": "^3.9.2",
     "@sebgroup/bootstrap": "^3.4.0",
     "@types/marked": "^0.6.5",
     "core-js": "^2.5.4",
-    "marked": "^0.6.2",
+    "marked": "^0.7.0",
     "prismjs": "^1.16.0",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.5.3",
     "tslib": "^1.9.0",
     "typescript-parser": "^2.6.1",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular-devkit/build-ng-packagr": "~0.13.0",
-    "@angular/cli": "~7.3.1",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
+    "@angular-devkit/build-angular": "~0.803.6",
+    "@angular-devkit/build-ng-packagr": "~0.803.6",
+    "@angular/cli": "~8.3.6",
+    "@angular/compiler-cli": "~8.2.8",
+    "@angular/language-service": "~8.2.8",
     "@commitlint/cli": "^7.5.2",
     "@commitlint/config-conventional": "^7.5.0",
     "@datorama/akita-ng-router-store": "^3.0.0",
@@ -85,7 +85,7 @@
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^8.9.5",
     "angular-exemplify": "^3.1.4",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "commitizen": "^3.1.1",
     "copyfiles": "^2.1.0",
     "cz-conventional-changelog": "^2.1.0",
@@ -101,7 +101,7 @@
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "mustache": "^3.0.1",
-    "ng-packagr": "^4.7.0",
+    "ng-packagr": "^5.4.0",
     "npm-run-all": "^4.1.5",
     "package-lock-sanitizer": "^1.0.0",
     "prettier": "^1.17.1",
@@ -111,9 +111,9 @@
     "semantic-release": "^16.0.0-beta.18",
     "travis-deploy-once": "^5.0.11",
     "ts-node": "~7.0.0",
-    "tsickle": ">=0.34.3",
+    "tsickle": "^0.37.0",
     "tslint": "~5.11.0",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "~3.2.2"
+    "typescript": "^3.5.3"
   }
 }
