@@ -1,21 +1,20 @@
 import { Component, OnInit } from '@angular/core';
-import {from, merge, Observable} from 'rxjs';
-import {ActivatedRoute} from '@angular/router';
-import {Declaration, TypescriptParser} from 'typescript-parser';
-import {File as ParsedFile} from 'typescript-parser/resources/File';
-import {ApiSection} from '../../../interfaces/api-section';
+import { from, merge, Observable } from 'rxjs';
+import { ActivatedRoute } from '@angular/router';
+import { Declaration, TypescriptParser } from 'typescript-parser';
+import { File as ParsedFile } from 'typescript-parser/resources/File';
+import { ApiSection } from '../../../interfaces/api-section';
 import marked from 'marked';
-import {map, reduce, tap} from 'rxjs/operators';
+import { map, reduce, tap } from 'rxjs/operators';
 
 @Component({
   selector: 'app-api-list',
   templateUrl: './api-list.component.html',
-  styleUrls: ['./api-list.component.scss']
+  styleUrls: ['./api-list.component.scss'],
 })
 export class ApiListComponent implements OnInit {
-
   $content: Observable<any>;
-  constructor(private route: ActivatedRoute) { }
+  constructor(private route: ActivatedRoute) {}
 
   ngOnInit() {
     const sources = this.route.routeConfig.data.sources;
@@ -23,7 +22,7 @@ export class ApiListComponent implements OnInit {
       return [...previous, this.parseSourceFile(current)];
     }, []);
     this.$content = merge(...obs).pipe(
-        reduce((previous, current) => [...previous, ...current],[]),
+      reduce(([previous, current]) => [...previous, ...current], [])
     );
   }
 
@@ -35,44 +34,58 @@ export class ApiListComponent implements OnInit {
     const methods = ApiListComponent.extractMethods(source);
     const tsParser = new TypescriptParser();
     return from(tsParser.parseSource(source)).pipe(
-        map(res => this.parse(res, description, inputs, outputs, properties, methods))
+      map(res =>
+        this.parse(res, description, inputs, outputs, properties, methods)
+      )
     );
   }
 
-  parse(file: ParsedFile, description: any, inputs: any, outputs: any, properties: any, methods: any): Array<ApiSection> {
-
+  parse(
+    file: ParsedFile,
+    description: any,
+    inputs: any,
+    outputs: any,
+    properties: any,
+    methods: any
+  ): Array<ApiSection> {
     return file.declarations
-      .filter(declaration => declaration.constructor.name === 'ClassDeclaration')
+      .filter(
+        declaration => declaration.constructor.name === 'ClassDeclaration'
+      )
       .reduce((previous, current: any) => {
-      const declaration: Declaration = current;
-      const section: ApiSection = {
-        description: description ? ApiListComponent.parseComment(description.groups.comment) : 'n/a',
-        name: declaration.name,
-        // @ts-ignore
-        inputs: this.parseInputs(declaration.accessors, inputs),
-        // @ts-ignore
-        outputs: this.parseOutputs(declaration.properties, outputs),
-        // @ts-ignore
-        properties: ApiListComponent.parseProperties(declaration.properties, properties),
-        // @ts-ignore
-        methods: ApiListComponent.parseMethods(declaration.methods, methods)
-      };
-      const isEmpty = !Object.entries(section)
+        const declaration: Declaration = current;
+        const section: ApiSection = {
+          description: description
+            ? ApiListComponent.parseComment(description.groups.comment)
+            : 'n/a',
+          name: declaration.name,
+          // @ts-ignore
+          inputs: this.parseInputs(declaration.accessors, inputs),
+          // @ts-ignore
+          outputs: this.parseOutputs(declaration.properties, outputs),
+          // @ts-ignore
+          properties: ApiListComponent.parseProperties(
+            declaration.properties,
+            properties
+          ),
+          // @ts-ignore
+          methods: ApiListComponent.parseMethods(declaration.methods, methods),
+        };
+        const isEmpty = !Object.entries(section)
           .filter(key => Array.isArray(key[1]))
           .some(key => key[1].length > 0);
-      return isEmpty ? [...previous] : [...previous, section];
-      }, []
-    );
+        return isEmpty ? [...previous] : [...previous, section];
+      }, []);
   }
 
   static extractInputs(sourceCode: string) {
     const regex = /(?<comment>\/\*\*(?:[\sA-Za-z\*\`\.\,\(\)\/\?\=\:\[\]\&\{\}]*)\*\/)?(?:[\r\n\t\s]*)(?<decorator>\@Input)\((?:'|"?)(?<alias>.*?)(?:'|")?(?:\))(?:[\W]+)(?<accessor>get|set|){1}(?:\W)?(?<name>[^\(]+)/g;
     let input = regex.exec(sourceCode);
-    let inputs = input ? {[input.groups.name] : input.groups} : {};
+    let inputs = input ? { [input.groups.name]: input.groups } : {};
     while (input !== null) {
       input = regex.exec(sourceCode);
       if (input) {
-        inputs = {...inputs, [input.groups.name] : input.groups};
+        inputs = { ...inputs, [input.groups.name]: input.groups };
       }
     }
     return inputs;
@@ -81,11 +94,11 @@ export class ApiListComponent implements OnInit {
   static extractOutputs(sourceCode: string) {
     const regex = /(?:\/\*\*(?<comment>[\s\S][^@]+)\*\/[^@]+|)(?<decorator>\@Output)\((?:'|"?)(?<alias>.*?)(?:'|")?(?:\))(?:\W)?(?<name>[^\:]+)/g;
     let output = regex.exec(sourceCode);
-    let outputs = output ? {[output.groups.name] : output.groups} : {};
+    let outputs = output ? { [output.groups.name]: output.groups } : {};
     while (output !== null) {
       output = regex.exec(sourceCode);
       if (output) {
-        outputs = {...outputs, [output.groups.name] : output.groups};
+        outputs = { ...outputs, [output.groups.name]: output.groups };
       }
     }
     return outputs;
@@ -94,11 +107,13 @@ export class ApiListComponent implements OnInit {
   static extractProperties(sourceCode: string) {
     const regex = /(?<name>[\w\$]+)\:\s(?<type>.[^\;\s]*)(?:\;\s| \=\s)[\'\"]?(?<default>[\w][^\;\/\'\"]*)?[\'\"]?(?:\;?\s?\/\/\s?(?<comment>.*))?/g;
     let property = regex.exec(sourceCode);
-    let properties = property ? {[property.groups.name] : property.groups} : {};
+    let properties = property
+      ? { [property.groups.name]: property.groups }
+      : {};
     while (property !== null) {
       property = regex.exec(sourceCode);
       if (property) {
-        properties = {...properties, [property.groups.name] : property.groups};
+        properties = { ...properties, [property.groups.name]: property.groups };
       }
     }
     return properties;
@@ -106,18 +121,18 @@ export class ApiListComponent implements OnInit {
   static extractMethods(sourceCode: string) {
     const regex = /(?:\/\*\*(?<comment>[\s\S][^\/]*)\*\/[^\w\@]+|)[^\w\]](?!constructor|Input|Component)(?<name>[a-z]*)\((?<parameters>[^\)]*)\)\:?\s?(?<returns>[\w\<\>]*)/g;
     let method = regex.exec(sourceCode);
-    let methods = method ? {[method.groups.name] : method.groups} : {};
+    let methods = method ? { [method.groups.name]: method.groups } : {};
     while (method !== null) {
       method = regex.exec(sourceCode);
       if (method) {
-        methods = {...methods, [method.groups.name] : method.groups};
+        methods = { ...methods, [method.groups.name]: method.groups };
       }
     }
     return methods;
   }
   static extractDescription(sourceCode: string) {
     const regex = /(?:\/\*\*(?<comment>[\s\S][^\/]*)\*\/[^\w])/;
-    console.log(regex.exec(sourceCode))
+    console.log(regex.exec(sourceCode));
     return regex.exec(sourceCode);
   }
 
@@ -130,64 +145,95 @@ export class ApiListComponent implements OnInit {
           input.type = current.type;
           return [...previous];
         }
-        return [...previous, {
-          ...current,
-          alias: inputs[current.name].alias,
-          description: ApiListComponent.parseComment(inputs[current.name].comment)
-        }];
+        return [
+          ...previous,
+          {
+            ...current,
+            alias: inputs[current.name].alias,
+            description: ApiListComponent.parseComment(
+              inputs[current.name].comment
+            ),
+          },
+        ];
       }, []);
   }
 
   parseOutputs(properties: Array<any>, outputs: any): Array<any> {
     return properties
-      .filter(property => property.type && property.type.indexOf('EventEmitter') !== -1)
-        .map(property => {
-          return {
-            ...property, description: ApiListComponent.parseComment(outputs[property.name].comment)
-          }
-        });
-  }
-
-  static parseProperties(properties: Array<any>, extractedProperties: any): Array<any> {
-    return properties
-      .filter(property =>
-        property.type && property.type.indexOf('EventEmitter') === -1 && // remove properties of type event emitter (Outputs)
-        property.name.substring(0, 1) !== '_' // remove private properties
+      .filter(
+        property =>
+          property.type && property.type.indexOf('EventEmitter') !== -1
       )
       .map(property => {
-        return extractedProperties[property.name] ? {
+        return {
           ...property,
-          default: extractedProperties[property.name].default,
-          type: extractedProperties[property.name].type,
-          description: ApiListComponent.parseComment(extractedProperties[property.name].comment)
-        } : {
-        ...property
+          description: ApiListComponent.parseComment(
+            outputs[property.name].comment
+          ),
         };
+      });
+  }
+
+  static parseProperties(
+    properties: Array<any>,
+    extractedProperties: any
+  ): Array<any> {
+    return properties
+      .filter(
+        property =>
+          property.type &&
+          property.type.indexOf('EventEmitter') === -1 && // remove properties of type event emitter (Outputs)
+          property.name.substring(0, 1) !== '_' // remove private properties
+      )
+      .map(property => {
+        return extractedProperties[property.name]
+          ? {
+              ...property,
+              default: extractedProperties[property.name].default,
+              type: extractedProperties[property.name].type,
+              description: ApiListComponent.parseComment(
+                extractedProperties[property.name].comment
+              ),
+            }
+          : {
+              ...property,
+            };
       });
   }
 
   static parseMethods(methods: Array<any>, extractedMethods: any): Array<any> {
     return methods
-      .filter(property =>
-        property.name.substring(0, 2) !== 'ng' && // remove angular lifecycle methods
-        property.name.substring(0, 1) !== '_') // remove private methods
+      .filter(
+        property =>
+          property.name.substring(0, 2) !== 'ng' && // remove angular lifecycle methods
+          property.name.substring(0, 1) !== '_'
+      ) // remove private methods
       .map(method => {
         return {
           ...method,
           functionCall:
-              method.name + '(' + (
-                  extractedMethods[method.name] ?
-                      extractedMethods[method.name].parameters :
-                      (method.parameters.map(param => param.name + ': ' + param.type).toString().replace(/,/g, ', '))) +
-              ')',
-          description: extractedMethods[method.name] ? ApiListComponent.parseComment(extractedMethods[method.name].comment) : 'n/a'
-
+            method.name +
+            '(' +
+            (extractedMethods[method.name]
+              ? extractedMethods[method.name].parameters
+              : method.parameters
+                  .map(param => param.name + ': ' + param.type)
+                  .toString()
+                  .replace(/,/g, ', ')) +
+            ')',
+          description: extractedMethods[method.name]
+            ? ApiListComponent.parseComment(
+                extractedMethods[method.name].comment
+              )
+            : 'n/a',
         };
       });
   }
 
   static parseComment(comment: string) {
-    return comment ? marked(comment.replace(/\*\/+|\/\*+|\*\s+|[\t\r\n]/g,'')) : 'n/a';
+    return comment
+      ? marked(comment.replace(/\*\/+|\/\*+|\*\s+|[\t\r\n]/g, ''))
+      : 'n/a';
   }
 
   static sortInputs(a, b) {
@@ -195,4 +241,3 @@ export class ApiListComponent implements OnInit {
     return isSet ? -1 : 0;
   }
 }
-
